Split analytics modules into separate components to isolate issues
